requirements.txt
algoliasearch-django>=2.0,<3.0
django>=4.0.0,<4.2.0
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
mysqlclient
dnspython
python-dateutil